Namespace compute integration tests
