--- integration/README.rst
@@ -1,20 +1,0 @@
-Integration Test Module
-=======================
-
-This test suite is for running a live API endpoint and testing the apache-libcloud functionality as a full integration test
-
-Running the API service
------------------------
-
-.. code-block:: bash
-
-  pip install -r integration/requirements.txt
-  python -m integration.api
-
-Running the tests
------------------
-
-.. code-block:: bash
-
-   python -m integration
-
