+++ integration/compute/README.rst
@@ -1,0 +1,20 @@
+Compute Integration Test Module
+===============================
+
+This test suite is for running a live API endpoint and testing the apache-libcloud compute functionality as a full integration test
+
+Running the API service
+-----------------------
+
+.. code-block:: bash
+
+  pip install -r integration/compute/requirements.txt
+  python -m integration.compute.api
+
+Running the tests
+-----------------
+
+.. code-block:: bash
+
+   python -m integration.compute
+
